+++ DesignToolbox/DesignToolbox/Pages/Components/Actions/Button/ButtonPage.swift
@@ -41,7 +41,6 @@ private struct ButtonDemo: View {
 
     @Environment(\.theme) private var theme
 
-    // swiftlint:disable accessibility_label_for_image
     var body: some View {
         HStack(alignment: .center) {
             Spacer()
@@ -52,7 +51,7 @@ private struct ButtonDemo: View {
             } else {
                 switch configurationModel.layout {
                 case .iconOnly:
-                    OUDSButton(icon: Image(systemName: "figure.handball"),
+                    OUDSButton(icon: Image.defaultImage(prefixedBy: theme.name),
                                accessibilityLabel: "app_components_common_icon_a11y".localized(),
                                flipIcon: configurationModel.flipIcon,
                                appearance: configurationModel.appearance,
@@ -63,7 +62,7 @@ private struct ButtonDemo: View {
                                style: configurationModel.style) {}
                 case .textAndIcon:
                     OUDSButton(text: configurationModel.text,
-                               icon: Image(systemName: "figure.handball"),
+                               icon: Image.defaultImage(prefixedBy: theme.name),
                                flipIcon: configurationModel.flipIcon,
                                appearance: configurationModel.appearance,
                                style: configurationModel.style) {}
@@ -75,5 +74,4 @@ private struct ButtonDemo: View {
         .disabled(!configurationModel.enabled)
         .padding(.all, theme.spaces.fixedMedium)
     }
-    // swiftlint:enable accessibility_label_for_image
 }