+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/CheckboxItem/Indeterminate/CheckboxItemIndeterminatePage.swift
@@ -57,10 +57,7 @@ private struct CheckboxItemIndeterminateDemo: View {
             .accessibilityIdentifier(A11YIdentifiers.componentCheckboxItemIndeterminate)
     }
 
-    // Need here that system name, a11y managed in component
-    // swiftlint:disable accessibility_label_for_image
     private var icon: Image? {
-        configurationModel.icon ? Image(systemName: "figure.handball") : nil
+        configurationModel.icon ? Image.defaultImage(prefixedBy: theme.name) : nil
     }
-    // swiftlint:enable accessibility_label_for_image
 }