+++ DesignToolbox/DesignToolbox/Pages/Components/Indicators/Tags/Tag/TagPage.swift
@@ -19,6 +19,7 @@ import SwiftUI
 struct TagPage: View {
 
     @StateObject private var configurationModel: TagConfigurationModel
+    @Environment(\.theme) private var theme
 
     init() {
         _configurationModel = StateObject(wrappedValue: TagConfigurationModel())
@@ -29,6 +30,8 @@ struct TagPage: View {
             TagDemo(configurationModel: configurationModel)
         } configurationView: {
             TagConfigurationView(configurationModel: configurationModel)
+        }.onAppear {
+            configurationModel.themeName = theme.name
         }
     }
 }